Rename module ed25519 -> ed25519wallet
