+++ packages/iov-keycontrol/src/keyring-entries/ed25519wallet.ts
