+++ packages/iov-keycontrol/src/keyring-entries/index.ts
@@ -1,3 +1,3 @@
-export * from "./ed25519";
+export { Ed25519Wallet } from "./ed25519wallet";
 export { Ed25519HdWallet } from "./ed25519hdwallet";
 export { Secp256k1HdWallet } from "./secp256k1hdwallet";
