+++ packages/iov-keycontrol/types/keyring-entries/ed25519wallet.d.ts
